Add Mainmenu. Major fix on screen resolution independence.
Regarding to screen resolution independence, use "apply()" or
"apply(true)" before begin to render content on screen in render loop;
depends on whether or not we need to center camera.

As it can be either hud, or cam, thus select a proper camera is always a
best practice to avoid headache in dealing with camera and viewport.

Drawing BitmapFont uses lower-left corner as origin for its coordinate
which is different from sprite coordinate whose origin is at the
top-left corner. See MenuItem class's render() method and its
initialization of boundingRect variable which are different in
y-position setting for example of this explaination.