Add hightlight as red for in-select of menu item. Fix lag of playing simulated bg music.
Playing simulated bg music is lagging as from testing due to .ogg file.
Guess it's because it needs to decompress the file everytime it needs to
play, and with the reason that the game plays such sound effects
frequently. Now they are converted into .wav file format. Tested and the
game runs smoothly.

Rendering BitmapFont with color needs to change its color at generation
time, not run time. It might be due to .ttf font itself. We cannot make
setColor() method of BitmapFont works.