Refactor batch render ability via interface class
IBatchShapeRenderable aims to be implemented as member function.

IBatchWrapperShapeRenderable aims to be implemented as static member
function (companion object) as there will be siuation that we need to
rener multiple of same type of object. But it's not required as user can
still use a single object to call its
beginBatchRender()/endBatchRender().