Refactor BBInput to support analog as digital key. Better use word.
Use controller not game pad when refer to in the code.
Controller is the bigger category, but game pad is sub-category.
Controller can include joystick. Thus this changes also include better
code fix for wording like those.

deadZone and sensitivity value is managed by BBInput. Sensitivity value
will be used with analog key processing.

This changes aim for a single-point of calling api as it would be
cumbersome to separate it into another set again.